+++ lib/apex_dropdown.dart
@@ -12,9 +12,11 @@ export 'src/models/chip_display.dart';
 export 'src/models/decoration.dart';
 
 export 'src/widgets/apex_async_dropdown.dart';
+export 'src/widgets/apex_async_multi_dropdown.dart';
 export 'src/widgets/apex_dropdown.dart';
 export 'src/widgets/apex_multi_dropdown.dart';
 
 export 'src/form_fields/apex_async_dropdown_form_field.dart';
+export 'src/form_fields/apex_async_multi_dropdown_form_field.dart';
 export 'src/form_fields/apex_dropdown_form_field.dart';
 export 'src/form_fields/apex_multi_dropdown_form_field.dart';